fix(input): onchange behaviour
