+++ src/styles/components/Main/index.style.ts
@@ -3,13 +3,14 @@ import styled from 'styled-components';;
 interface MainPropsType {
   height?: string
   background?: string
+  overflowx?: string
 }
 
 export const Main = styled.div<MainPropsType>`
   margin: 80px 0 0;
   height: ${(props) => props.height || '80vh'};
   width: 100%;
-  overflow-x: scroll;
+  overflow-x: ${(props) => props.overflowx || 'inherit'};
   background: ${(props) => props.background || 'inherit'};
   display: flex;
   flex-direction: column;
@@ -18,4 +19,7 @@ export const Main = styled.div<MainPropsType>`
 `;
 
 Main.shouldForwardProp =
-  (prop) => prop !== 'height' && prop !== 'background';
+  (prop) =>
+    prop !== 'height' &&
+    prop !== 'background' &&
+    prop !== 'overflow-x';
